make EmbedColors inner class and fix Nullable annotations
