+++ src/main/java/de/nplay/moderationbot/Helpers.java
@@ -2,7 +2,7 @@ package de.nplay.moderationbot;
 
 import com.github.kaktushose.jda.commands.dispatching.events.ReplyableEvent;
 import com.github.kaktushose.jda.commands.embeds.Embed;
-import de.nplay.moderationbot.embeds.EmbedColors;
+import de.nplay.moderationbot.NPLAYModerationBot.EmbedColors;
 import de.nplay.moderationbot.messagelink.MessageLink;
 import de.nplay.moderationbot.moderation.ModerationService.ModerationAct;
 import net.dv8tion.jda.api.entities.Message;
