--- src/main/java/de/nplay/moderationbot/embeds/EmbedColors.java
@@ -1,19 +1,0 @@
-package de.nplay.moderationbot.embeds;
-
-public enum EmbedColors {
-    DEFAULT("#020c24"),
-    ERROR("#ff0000"),
-    SUCCESS("#00ff00"),
-    WARNING("#ffff00");
-
-    public final String hexColor;
-
-    EmbedColors(String hexColor) {
-        this.hexColor = hexColor;
-    }
-
-    @Override
-    public String toString() {
-        return hexColor;
-    }
-}
